+++ app/assets/javascripts/govuk-admin-template/modules/auto_track_event.js
@@ -4,11 +4,12 @@
   Modules.AutoTrackEvent = function() {
     var that = this;
     that.start = function(element) {
-      var action = element.data('track-action'),
+      var category = element.data('track-category'),
+          action = element.data('track-action'),
           label = element.data('track-label'),
           value = element.data('track-value');
 
-      GOVUKAdmin.trackEvent(action, label, value);
+      GOVUKAdmin.trackEvent(category, action, { label: label, value: value });
     }
   };