+++ spec/javascripts/spec/auto_track_event.spec.js
@@ -7,7 +7,7 @@ describe('An auto event tracker', function() {
 
   beforeEach(function() {
     element = $('\
-      <div data-track-action="action" data-track-label="label" data-track-value="10">\
+      <div data-track-category="category" data-track-action="action" data-track-label="label" data-track-value="10">\
       </div>\
     ');
 
@@ -17,6 +17,6 @@ describe('An auto event tracker', function() {
   it('tracks events on start', function() {
     spyOn(root.GOVUKAdmin, 'trackEvent');
     tracker.start(element);
-    expect(GOVUKAdmin.trackEvent).toHaveBeenCalledWith('action', 'label', 10)
+    expect(GOVUKAdmin.trackEvent).toHaveBeenCalledWith('category', 'action', {label: 'label', value: 10 })
   });
 });